+++ src/services/TyreModelService.js
@@ -14,6 +14,16 @@ export default {
   create(dto) {
     return axios.post(`${baseURL}`, dto)
   },
+  uploadImage(id, file) {
+    const formData = new FormData()
+    formData.append('file', file)
+
+    return axios.post(`${baseURL}/${id}/upload`, formData, {
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      }
+    })
+  },
   update(id, dto) {
     return axios.patch(`${baseURL}/${id}`, dto)
   },